+++ uikit-sample/src/main/java/com/sendbird/uikit_messaging_android/BaseApplication.java
@@ -1,7 +1,6 @@
 package com.sendbird.uikit_messaging_android;
 
-
-import android.app.Application;
+import androidx.multidex.MultiDexApplication;
 
 import com.sendbird.uikit.SendBirdUIKit;
 import com.sendbird.uikit.adapter.SendBirdUIKitAdapter;
@@ -10,7 +9,7 @@ import com.sendbird.uikit_messaging_android.fcm.MyFirebaseMessagingService;
 import com.sendbird.uikit_messaging_android.utils.PreferenceUtils;
 import com.sendbird.uikit_messaging_android.utils.PushUtils;
 
-public class BaseApplication extends Application {
+public class BaseApplication extends MultiDexApplication {
 
     private static final String APP_ID = "2D7B4CDB-932F-4082-9B09-A1153792DC8D";