+++ uikit/src/main/java/com/sendbird/uikit/fragments/ChannelFragment.java
@@ -7,6 +7,7 @@ import android.content.Context;
 import android.content.Intent;
 import android.content.res.ColorStateList;
 import android.net.Uri;
+import android.os.Build;
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -15,6 +16,7 @@ import android.view.ViewGroup;
 import androidx.annotation.DrawableRes;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
+import androidx.annotation.RequiresApi;
 import androidx.annotation.StringRes;
 import androidx.annotation.StyleRes;
 import androidx.databinding.DataBindingUtil;
@@ -680,7 +682,13 @@ public class ChannelFragment extends BaseGroupChannelFragment implements OnItemC
         checkPermission(PERMISSION_REQUEST_ALL, new IPermissionHandler() {
             @Override
             public String[] getPermissions(int requestCode) {
-                return new String[]{Manifest.permission.CAMERA, Manifest.permission.WRITE_EXTERNAL_STORAGE, Manifest.permission.READ_EXTERNAL_STORAGE};
+                if (Build.VERSION.SDK_INT > Build.VERSION_CODES.P) {
+                    return new String[]{Manifest.permission.CAMERA,
+                            Manifest.permission.READ_EXTERNAL_STORAGE};
+                }
+                return new String[]{Manifest.permission.CAMERA,
+                        Manifest.permission.WRITE_EXTERNAL_STORAGE,
+                        Manifest.permission.READ_EXTERNAL_STORAGE};
             }
 
             @Override
@@ -704,7 +712,11 @@ public class ChannelFragment extends BaseGroupChannelFragment implements OnItemC
         checkPermission(PERMISSION_REQUEST_STORAGE, new IPermissionHandler() {
             @Override
             public String[] getPermissions(int requestCode) {
-                return new String[]{Manifest.permission.WRITE_EXTERNAL_STORAGE, Manifest.permission.READ_EXTERNAL_STORAGE};
+                if (Build.VERSION.SDK_INT > Build.VERSION_CODES.P) {
+                    return new String[]{Manifest.permission.READ_EXTERNAL_STORAGE};
+                }
+                return new String[]{Manifest.permission.WRITE_EXTERNAL_STORAGE,
+                        Manifest.permission.READ_EXTERNAL_STORAGE};
             }
 
             @Override
@@ -725,7 +737,11 @@ public class ChannelFragment extends BaseGroupChannelFragment implements OnItemC
         checkPermission(PERMISSION_REQUEST_STORAGE, new IPermissionHandler() {
             @Override
             public String[] getPermissions(int requestCode) {
-                return new String[]{Manifest.permission.WRITE_EXTERNAL_STORAGE, Manifest.permission.READ_EXTERNAL_STORAGE};
+                if (Build.VERSION.SDK_INT > Build.VERSION_CODES.P) {
+                    return new String[]{Manifest.permission.READ_EXTERNAL_STORAGE};
+                }
+                return new String[]{Manifest.permission.WRITE_EXTERNAL_STORAGE,
+                        Manifest.permission.READ_EXTERNAL_STORAGE};
             }
 
             @Override
@@ -1161,27 +1177,34 @@ public class ChannelFragment extends BaseGroupChannelFragment implements OnItemC
             } else if (key == R.string.sb_text_channel_anchor_delete) {
                 showWarningDialog(message);
             } else if (key == R.string.sb_text_channel_anchor_save) {
-                checkPermission(PERMISSION_REQUEST_STORAGE, new IPermissionHandler() {
-                    @Override
-                    public String[] getPermissions(int requestCode) {
-                        return new String[]{Manifest.permission.READ_EXTERNAL_STORAGE, Manifest.permission.WRITE_EXTERNAL_STORAGE};
-                    }
+                if (Build.VERSION.SDK_INT > Build.VERSION_CODES.P) {
+                    download((FileMessage) message);
+                } else {
+                    checkPermission(PERMISSION_REQUEST_STORAGE, new IPermissionHandler() {
+                        @Override
+                        public String[] getPermissions(int requestCode) {
+                            return new String[]{Manifest.permission.WRITE_EXTERNAL_STORAGE,
+                                    Manifest.permission.READ_EXTERNAL_STORAGE};
+                        }
 
-                    @Override
-                    public void onPermissionGranted(int requestCode) {
-                        download((FileMessage) message);
-                    }
-                });
+                        @Override
+                        public void onPermissionGranted(int requestCode) {
+                            download((FileMessage) message);
+                        }
+                    });
+                }
             }
         };
     }
 
-    private void download(FileMessage message) {
+    private void download(@NonNull FileMessage fileMessage) {
         toastSuccess(R.string.sb_text_toast_success_start_download_file);
         TaskQueue.addTask(new JobResultTask<Boolean>() {
+            @RequiresApi(api = Build.VERSION_CODES.O)
             @Override
             public Boolean call() throws Exception {
-                FileDownloader.getInstance().saveFile(getContext(), message.getUrl(), message.getName());
+                FileDownloader.getInstance().saveFile(getContext(), fileMessage.getUrl(),
+                        fileMessage.getType(), fileMessage.getName());
                 return true;
             }