+++ uikit/src/main/java/com/sendbird/uikit/fragments/UserTypeListFragment.java
@@ -195,7 +195,29 @@ abstract public class UserTypeListFragment extends BaseGroupChannelFragment impl
             Logger.dev("++ observing result users size : %s", users.size());
             adapter.setItems((List<User>) users, channel.getMyRole());
         });
+        viewModel.getOperatorDismissed().observe(this, isDismissed -> {
+            if (isDismissed) onOperatorDismissed();
+        });
+        viewModel.getChannelDeleted().observe(this, isDeleted -> {
+            if (isDeleted) onChannelDeleted();
+        });
         viewModel.loadInitial();
+    }
+
+    /**
+     * Called when a user is dismissed from the operator.
+     *
+     * @since 2.1.3
+     */
+    protected void onOperatorDismissed() {}
+
+    /**
+     * Called when a channel is deleted from the user.
+     *
+     * @since 2.1.3
+     */
+    protected void onChannelDeleted() {
+        finish();
     }
 
     /**