+++ uikit/src/main/java/com/sendbird/uikit/vm/OpenChannelViewModel.java
@@ -38,8 +38,8 @@ import java.util.concurrent.atomic.AtomicReference;
 
 public class OpenChannelViewModel extends BaseViewModel implements LifecycleObserver, PagerRecyclerView.Pageable<List<BaseMessage>> {
     private static final int DEFAULT_MESSAGE_LOAD_SIZE = 40;
-    private static final String CONNECTION_HANDLER_ID = "CONNECTION_HANDLER_GROUP_CHAT";
-    private static final String CHANNEL_HANDLER_ID = "CHANNEL_HANDLER_GROUP_CHANNEL_CHAT";
+    private final String CONNECTION_HANDLER_ID = "CONNECTION_HANDLER_OPEN_CHAT" + System.currentTimeMillis();;
+    private final String CHANNEL_HANDLER_ID = "CHANNEL_HANDLER_OPEN_CHANNEL_CHAT" + System.currentTimeMillis();;
     private final ExecutorService worker = Executors.newSingleThreadExecutor();
     private final MutableLiveData<List<BaseMessage>> messageList = new MutableLiveData<>();
     private final ChatMessageCollection messageCollection = new ChatMessageCollection();
@@ -96,8 +96,8 @@ public class OpenChannelViewModel extends BaseViewModel implements LifecycleObse
         return messageLoadState;
     }
 
-    @OnLifecycleEvent(Lifecycle.Event.ON_RESUME)
-    private void onResume() {
+    @OnLifecycleEvent(Lifecycle.Event.ON_CREATE)
+    private void onCreate() {
         SendBird.addConnectionHandler(CONNECTION_HANDLER_ID, new SendBird.ConnectionHandler() {
             @Override
             public void onReconnectStarted() {
@@ -267,7 +267,10 @@ public class OpenChannelViewModel extends BaseViewModel implements LifecycleObse
                 }
             }
         });
+    }
 
+    @OnLifecycleEvent(Lifecycle.Event.ON_RESUME)
+    private void onResume() {
         requestChangeLogs(channel);
     }